requirements.txt
aiodns==3.2.0
aiohttp==3.9.3
aiosignal==1.3.1
attrs==23.2.0
ccxt==4.2.89
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==42.0.5
frozenlist==1.4.1
idna==3.6
markdown-it-py==3.0.0
mdurl==0.1.2
multidict==6.0.5
numpy==1.26.4
pandas==2.2.1
pycares==4.4.0
pycparser==2.22
Pygments==2.17.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
requests==2.31.0
rich==13.7.1
six==1.16.0
ta==0.11.0
transitions==0.9.0
typing_extensions==4.10.0
tzdata==2024.1
urllib3==2.2.1
yarl==1.9.4
